change the splash screen with flutter native splash
