+++ android/app/src/main/res/drawable-mdpi/android12splash.png
