+++ android/app/src/main/res/drawable-night/launch_background.xml
