+++ android/app/src/main/res/drawable-xxhdpi/android12splash.png
