+++ lib/my_application.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+
+class MyApplication extends StatelessWidget {
+  const MyApplication({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const Scaffold(
+        body: Center(child: Text("Hello"),),
+      ),
+    );
+  }
+}
+
